refactored raytracing

+++ src/Core/level/room/Point.java
@@ -1,27 +1,27 @@
 package core.level.room;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void moveX(int x) {
+    public void moveX(double x) {
         this.x += x;
     }
 
-    public void moveY(int y) {
+    public void moveY(double y) {
         this.y += y;
     }
 }
